py-sidecars/requirements.txt
paddleocr
paddlepaddle
pymupdf
opencv-python
py7zr
rarfile
Pillow
docx2pdf
pyinstaller
